requirements.txt
web.py
mysql-connector-python